pyproject.toml
[project]
name = "rdfcraft-python"
version = "0.1.0"
description = "A Python project for RDF crafting"
authors = [
    { name = "Ensar Emir EROL", email = "ensar.erol@maastrichtuniversity.nl" }
]
requires-python = "==3.11.*"

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asgi-correlation-id==4.3.4",
    "bottle==0.13.2",
    "bytecode==0.16.0",
    "certifi==2024.8.30",
    "click==8.1.7",
    "ddtrace==2.16.0",
    "Deprecated==1.2.14",
    "dnspython==2.7.0",
    "email_validator==2.2.0",
    "envier==0.6.1",
    "fastapi==0.115.2",
    "fastapi-cli==0.0.5",
    "h11==0.14.0",
    "html5lib-modern==1.2",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "importlib_metadata==8.5.0",
    "Jinja2==3.1.4",
    "kink==0.8.1",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "mdurl==0.1.2",
    "nuitka>=2.5.4",
    "numpy==2.1.2",
    "opentelemetry-api==1.28.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "protobuf==5.28.3",
    "proxy_tools==0.1.0",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "pygments==2.18.0",
    "pyobjc-core==10.3.1; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa==10.3.1; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz==10.3.1; sys_platform == 'darwin'",
    "pyobjc-framework-Security==10.3.1; sys_platform == 'darwin'",
    "pyobjc-framework-WebKit==10.3.1; sys_platform == 'darwin'",
    "pyparsing==3.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "pytz==2024.2",
    "pywebview==5.3.1",
    "PyYAML==6.0.2",
    "rdflib==7.1.0",
    "rich==13.9.2",
    "setuptools==70.3.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.36",
    "starlette==0.40.0",
    "structlog==24.4.0",
    "typer==0.12.5",
    "typing_extensions==4.12.2",
    "tzdata==2024.2",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
    "wrapt==1.16.0",
    "xmltodict==0.14.2",
    "zipp==3.20.2",
]

[dependency-groups]
dev = [
    "coverage>=7.6.8",
    "ruff>=0.8.0",
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "from .* import .*",
    "import .*",
    "return self.__repr__\\(\\)",
    ]
